chore: swapdetailsinput refactored into multiple files
