+++ packages/cysync-core/src/pages/MainApp/Swap/Pages/SwapDetailsInput.tsx
@@ -1,435 +1,29 @@
-// TODO: refactor this file into multiple components
-import {
-  formatDisplayPrice,
-  getParsedAmount,
-  getDefaultUnit,
-} from '@cypherock/coin-support-utils';
+import { getParsedAmount, getDefaultUnit } from '@cypherock/coin-support-utils';
 import { ServerErrorType } from '@cypherock/cysync-core-constants';
 import {
   Container,
   LangDisplay,
   Button,
   Flex,
-  Image,
-  Dropdown,
-  Input,
   Typography,
   VectorIcon,
-  CustomInputSend,
   Throbber,
-  ClockIcon,
   parseLangTemplate,
-  Chip,
 } from '@cypherock/cysync-ui';
-import { BigNumber, formatSecondsToMinutes } from '@cypherock/cysync-utils';
+import { BigNumber } from '@cypherock/cysync-utils';
 import { IAccount } from '@cypherock/db-interfaces';
 import lodash from 'lodash';
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
-import { IQuote, useSwap } from '~/context';
+import { IQuote } from '~/context';
 import { useAccountDropdown, useWalletDropdown } from '~/hooks';
 import { getQuotes } from '~/services/swapService';
-import { useAppSelector, selectLanguage, selectPriceInfos } from '~/store';
+import { useAppSelector, selectLanguage } from '~/store';
 import { createServerErrorFromError } from '~/utils';
 import logger from '~/utils/logger';
+import { AmountAndAccountSelection, SwapQuotes } from '../components';
 
 const throbber: JSX.Element = <Throbber size={15} strokeWidth={2} />;
-
-const AmountInput: React.FC<any> = ({
-  placeholder,
-  amount,
-  isDisabled,
-  isLoading,
-  coinValue,
-  coinUnit,
-  setAmount,
-  error,
-}) => {
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  const filterNumericInput = (val: string) => {
-    let filteredValue = val.replace(/[^0-9.]/g, '');
-    const bigNum = new BigNumber(filteredValue);
-
-    if (filteredValue.includes('.')) {
-      const splitValue = filteredValue.split('.');
-      let firstValue = splitValue[0];
-      const secondValue = splitValue[1];
-
-      const firstValBigNumber = new BigNumber(firstValue);
-
-      if (firstValBigNumber.isNaN() || firstValBigNumber.isZero()) {
-        firstValue = '0';
-      }
-
-      filteredValue = `${firstValue}.${secondValue}`;
-    } else if (!bigNum.isNaN() && bigNum.isZero()) {
-      filteredValue = '0';
-    }
-
-    return filteredValue;
-  };
-
-  const adjustInputCursor = () => {
-    if (
-      inputRef.current?.selectionEnd &&
-      inputRef.current.selectionEnd > amount.toString().length
-    ) {
-      inputRef.current.setSelectionRange(amount.length, amount.length);
-    }
-  };
-
-  const handleCoinAmountChange = (val: string) => {
-    const filteredValue = filterNumericInput(val);
-    setAmount(filteredValue);
-  };
-
-  return (
-    <>
-      <CustomInputSend error={error}>
-        <Input
-          ref={inputRef}
-          type="text"
-          name="amount"
-          placeholder={placeholder}
-          onChange={handleCoinAmountChange}
-          value={`${amount} ${coinUnit}`}
-          onClick={adjustInputCursor}
-          onKeyDown={e => e.code === 'Backspace' && adjustInputCursor()}
-          disabled={isDisabled}
-          $textColor="white"
-          $noBorder
-        />
-        {isLoading ? (
-          throbber
-        ) : (
-          <Typography
-            $fontSize={16}
-            color="muted"
-            $allowOverflow
-            $whiteSpace="nowrap"
-          >
-            {coinValue}
-          </Typography>
-        )}
-      </CustomInputSend>
-      {error && (
-        <Typography
-          variant="span"
-          color="error"
-          $alignSelf="start"
-          $fontSize={12}
-        >
-          {error}
-        </Typography>
-      )}
-    </>
-  );
-};
-
-const AmountAndAccountSelection: React.FC<any> = ({
-  selectionLabel,
-  amountLabel,
-  accountPlaceholder,
-  selectedWallet,
-  handleWalletChange,
-  walletDropdownList,
-  selectedAccount,
-  handleAccountChange,
-  accountDropdownList,
-  amount,
-  setAmount,
-  isAmountDisabled,
-  amountError,
-}) => {
-  const lang = useAppSelector(selectLanguage);
-  const dialogText = lang.strings.swap.detailsInput.common;
-
-  const { priceInfos } = useAppSelector(selectPriceInfos);
-
-  const coinUnit = useMemo(() => {
-    const account = selectedAccount;
-
-    if (!account) return '';
-
-    const unit =
-      account.unit ??
-      getDefaultUnit(account.parentAssetId, account.assetId).abbr;
-    return unit;
-  }, [selectedAccount]);
-
-  const coinValue = useMemo(() => {
-    const account = selectedAccount;
-    if (!account) return '';
-
-    const assetPrice = priceInfos.find(
-      p => p.assetId === account?.assetId && p.currency.toLowerCase() === 'usd',
-    );
-
-    if (!assetPrice) return '';
-
-    const validAmount = amount ?? '0';
-    const amountValue = new BigNumber(validAmount).multipliedBy(
-      assetPrice.latestPrice,
-    );
-
-    const value = formatDisplayPrice(amountValue);
-
-    return `~${value} ${assetPrice.currency.toUpperCase()}`;
-  }, [selectedAccount, amount]);
-
-  return (
-    <Flex direction="column" gap={16} $minWidth="420px">
-      <Flex direction="column" gap={8}>
-        <Typography $fontSize={12} color="muted">
-          {selectionLabel}
-        </Typography>
-        <Dropdown
-          items={walletDropdownList}
-          selectedItem={selectedWallet?.__id}
-          searchText={dialogText.searchText}
-          placeholderText={dialogText.walletPlaceholder}
-          onChange={handleWalletChange}
-          noLeftImageInList
-        />
-        <Dropdown
-          items={accountDropdownList}
-          selectedItem={selectedAccount?.__id}
-          disabled={!selectedWallet}
-          searchText={dialogText.searchText}
-          placeholderText={accountPlaceholder}
-          onChange={handleAccountChange}
-        />
-      </Flex>
-      <Flex direction="column" gap={8}>
-        <Flex align="center" justify="space-between">
-          <Typography $fontSize={12} color="muted">
-            {amountLabel}
-          </Typography>
-        </Flex>
-        <AmountInput
-          placeholder="0"
-          amount={amount}
-          setAmount={setAmount}
-          isDisabled={!selectedAccount || isAmountDisabled}
-          error={amountError}
-          coinValue={coinValue}
-          coinUnit={coinUnit}
-        />
-      </Flex>
-    </Flex>
-  );
-};
-
-const OfferBox: React.FC<any> = ({
-  selectedIndex,
-  setSelectedIndex,
-  offerData,
-}) => (
-  <Flex
-    p="20px"
-    px={2}
-    gap={16}
-    direction="column"
-    $borderRadius={8}
-    $borderColor={selectedIndex === offerData.index ? 'gold' : 'card'}
-    $borderWidth={1}
-    onClick={() => {
-      setSelectedIndex(offerData.index);
-    }}
-  >
-    <Flex justify="space-between" align="center">
-      <Flex gap={8} align="center">
-        <Image
-          src={offerData?.provider?.imageUrl ?? ''}
-          alt="Logo"
-          $width={32}
-          $height={32}
-        />
-        <Typography $fontSize={14}>{offerData.provider.name}</Typography>
-      </Flex>
-      {offerData.isBest && (
-        <Chip
-          $gradient="golden"
-          $fontSize={10}
-          $fontWeight="semibold"
-          px="8px"
-          py="5px"
-        >
-          {offerData.bestOfferText}
-        </Chip>
-      )}
-    </Flex>
-    <Flex direction="column">
-      {offerData?.data?.map((data: any) => (
-        <Flex justify="space-between" align="center" key={data.title}>
-          <Flex gap={8} align="center">
-            <Typography $fontSize={14}>{data.title}</Typography>
-          </Flex>
-          <Flex gap={8} align="center">
-            <Typography $fontSize={14}>{data.value[0]}</Typography>
-            <Typography $fontSize={12} color="muted">
-              {data.value[1]}
-            </Typography>
-          </Flex>
-        </Flex>
-      ))}
-    </Flex>
-  </Flex>
-);
-
-export const SwapQuotesHeader: React.FC<{
-  size: number;
-  onTimeEnd: () => void;
-}> = ({ size, onTimeEnd }) => {
-  const [seconds, setSeconds] = useState(30);
-
-  const lang = useAppSelector(selectLanguage);
-  const displayText = lang.strings.swap.detailsInput.offers;
-
-  useEffect(() => {
-    const interval = setInterval(
-      () => setSeconds(s => Math.max(s - 1, 0)),
-      1000,
-    );
-    return () => clearInterval(interval);
-  }, []);
-
-  useEffect(() => {
-    if (seconds <= 1) onTimeEnd();
-  }, [seconds]);
-
-  const remainingTime = useMemo(
-    () => formatSecondsToMinutes(seconds),
-    [seconds],
-  );
-
-  return (
-    <Typography
-      color="muted"
-      justify="space-between"
-      display="flex"
-      $allowOverflow
-    >
-      <span>{parseLangTemplate(displayText.quotesFound, { num: size })}</span>
-      <Typography color="muted" align="center" display="flex" gap={8}>
-        {parseLangTemplate(displayText.timerText)}
-        <Typography
-          color="white"
-          align="center"
-          display="flex"
-          gap={8}
-          $minWidth="71px"
-        >
-          <ClockIcon />
-          {remainingTime}
-        </Typography>
-      </Typography>
-    </Typography>
-  );
-};
-
-export const SwapQuotes: React.FC<{
-  quotes: IQuote[];
-  setSelectedOfferIndex: (val: number) => void;
-  selectedOfferIndex?: number;
-  toAccount?: IAccount;
-  fromAccount?: IAccount;
-  findNewQuotes: () => void;
-  hasEnoughBalance: boolean;
-}> = ({
-  quotes,
-  setSelectedOfferIndex,
-  selectedOfferIndex,
-  toAccount,
-  fromAccount,
-  findNewQuotes,
-  hasEnoughBalance,
-}) => {
-  const { toNextPage, fillDetails } = useSwap();
-  const lang = useAppSelector(selectLanguage);
-  const displayText = lang.strings.swap.detailsInput.offers;
-
-  return (
-    <>
-      <SwapQuotesHeader size={quotes.length} onTimeEnd={findNewQuotes} />
-      {quotes.map((quote, index) => {
-        const fromUnit = getDefaultUnit(
-          fromAccount?.parentAssetId ?? '',
-          fromAccount?.assetId,
-        ).abbr;
-
-        const toUnit = getDefaultUnit(
-          toAccount?.parentAssetId ?? '',
-          toAccount?.assetId,
-        ).abbr;
-
-        const rate = new BigNumber(quote.toAmount)
-          .dividedBy(quote.fromAmount)
-          .toFixed(6);
-
-        return (
-          <OfferBox
-            offerData={{
-              index,
-              data: [
-                {
-                  title: displayText.fixedRate,
-                  value: [`1 ${fromUnit} =`, `${rate} ${toUnit}`],
-                },
-              ],
-              isBest: index === 0,
-              bestOfferText: displayText.bestOffer,
-              provider: quote.provider,
-            }}
-            setSelectedIndex={setSelectedOfferIndex}
-            selectedIndex={selectedOfferIndex}
-            key={quote.id}
-          />
-        );
-      })}
-      {quotes.length === 0 && displayText.errors.noQuotes}
-      <Button
-        variant="primary"
-        disabled={
-          selectedOfferIndex === undefined ||
-          selectedOfferIndex >= quotes.length ||
-          !hasEnoughBalance
-        }
-        onClick={() => {
-          if (
-            !fromAccount ||
-            !toAccount ||
-            (selectedOfferIndex ?? 0) >= quotes.length
-          ) {
-            return;
-          }
-          const quote = quotes[selectedOfferIndex ?? 0];
-          fillDetails({
-            from: fromAccount,
-            to: toAccount,
-            quote,
-          });
-          toNextPage();
-        }}
-        display="inline-block"
-      >
-        <LangDisplay text={displayText.buttons.continue} />
-      </Button>
-    </>
-  );
-};
-
-SwapQuotes.defaultProps = {
-  selectedOfferIndex: undefined,
-  toAccount: undefined,
-  fromAccount: undefined,
-};
 
 export const SwapDetailsInput = () => {
   const [selectedOfferIndex, setSelectedOfferIndex] = useState<
